requirements.txt
streamlit
llama-index
llama-index-llms-ollama
llama-index-llms-gemini
llama-index-embeddings-huggingface
sentence-transformers
pypdf
python-dotenv